requirements.txt
google-genai
Flask